# libs/client-codegen/specs/core-api-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.0.1",
    "title": "Blameable Core API"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Base Path"
    }
  ],
...
  "paths": {
...
    "/admin/sites": {
      "get": {
        "description": "lists the sites for an authenticated user",
        "responses": {
          "200": {
            "description": "successfully retrieved site theme",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "isOwner": {
                        "type": "boolean"
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "isOwner"
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "unable to retrieve the sites"
          }
        }
      },
      "post": {
        "description": "creates a new site for an authenticated user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "defaultLocale": {
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "successfully created the site",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "defaultLocale": {
                      "type": "string"
                    },
                    "isOwner": {
                      "type": "boolean"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "defaultLocale",
                    "isOwner",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "unable to create the site"
          }
        }
      }
    },
    "/admin/sites/{siteId}": {
      "get": {
        "description": "gets a site details for an authenticated user",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "afskhfkjs"
            },
            "required": true,
            "name": "siteId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "successfully retrieved the site",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "defaultLocale": {
                      "type": "string"
                    },
                    "isOwner": {
                      "type": "boolean"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "defaultLocale",
                    "isOwner",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "unable to retrieve the site"
          }
        }
      },
      "patch": {
        "description": "patch a site for an authenticated user",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "afskhfkjs"
            },
            "required": true,
            "name": "siteId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "defaultLocale": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "site was deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "defaultLocale": {
                      "type": "string"
                    },
                    "isOwner": {
                      "type": "boolean"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "defaultLocale",
                    "isOwner",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "unable to delete the site"
          }
        }
      },
      "delete": {
        "description": "deletes a site for an authenticated user",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "afskhfkjs"
            },
            "required": true,
            "name": "siteId",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "site was deleted successfully"
          },
          "400": {
            "description": "unable to delete the site"
          }
        }
      }
    },
...
  }
}